add tests, split to analysis part and transformation
