--- packages/codemods/react/19/remove-legacy-context/__testfixtures__/class-component.ouput.tsx
@@ -1,19 +1,0 @@
-const FooContext = React.createContext();
-
-class Parent extends React.Component {
-  render() {
-    return (
-      <FooContext value='bar'>
-        <Child />
-      </FooContext>
-    );
-  }
-}
-
-class Child extends React.Component {
-  static contextType = FooContext;
-
-  render() {
-    return <div>{this.context}</div>;
-  }
-}
