+++ packages/codemods/react/19/remove-legacy-context/__testfixtures__/class-component.output.tsx
@@ -1,0 +1,10 @@
+import PropTypes from 'prop-types';
+import React from 'react';
+
+const Context = React.createContext();
+
+class Parent extends React.Component {
+  render() {
+    return <Context value={{ foo: 'bar' }}><Child /></Context>;
+  }
+}
