--- packages/codemods/react/19/remove-legacy-context/__testfixtures__/fixture1.input.ts
@@ -1,1 +1,0 @@
-const toReplace = "hello";
