--- packages/codemods/react/19/remove-legacy-context/__testfixtures__/fixture1.output.ts
@@ -1,1 +1,0 @@
-const replacement = "hello";
